refactor(data_prep): reorganize data processing modules
