--- src/_data_prep/README.txt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:81dd2a214f99536c5bbc7e51de66b4e45177094a3469e3f038d7d659d7f5b78b
-size 25
